ai_service/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
google-generativeai>=0.4.1
asyncpg>=0.29.0
